+++ griptape/tools/__init__.py
@@ -26,6 +26,8 @@ from .outpainting_image_generation_client.tool import OutpaintingImageGeneration
 from .griptape_cloud_knowledge_base_client.tool import GriptapeCloudKnowledgeBaseClient
 from .structure_run_client.tool import StructureRunClient
 from .image_query_client.tool import ImageQueryClient
+from .text_to_speech_client.tool import TextToSpeechClient
+from .audio_transcription_client.tool import AudioTranscriptionClient
 
 __all__ = [
     "BaseTool",
@@ -56,4 +58,6 @@ __all__ = [
     "GriptapeCloudKnowledgeBaseClient",
     "StructureRunClient",
     "ImageQueryClient",
+    "TextToSpeechClient",
+    "AudioTranscriptionClient",
 ]